pyproject.toml
[project]
name = "onepromptai"
version = "0.1.0"
description = "Autonomous multi-agent coding orchestrator — swarm system for building projects from a single prompt"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "boto3>=1.34.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "gitpython>=3.1.40",
    "click>=8.1.7",
    "pydantic>=2.5.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
mongodb = ["pymongo>=4.6.0"]

[project.scripts]
onepromptai = "main:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
